refactor: move kubernetes part to its own file
so it'll be easier to work on adding two tabs only for Kubernetes part

related to https://github.com/kortex-hub/kortex/issues/281
